All functionality works

+++ app/Forum/Category/Controllers/DeleteController.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace App\Forum\Category\Controllers;
 
+use App\Access\Models\AccessChecker\Forum\CategoryAccessChecker;
+use App\Access\Models\Forbidden;
 use App\Forum\Category\Models\CategoryRepository;
 use Ramsey\Uuid\Uuid;
 
@@ -11,12 +13,23 @@ final class DeleteController extends \Phalcon\Mvc\Controller
 {
     public function mainAction(string $id): void
     {
+        if (!$this->getCategoryAccessChecker()->canDelete()) {
+            throw new Forbidden();
+        }
+
         $category = $this->getCategoryRepository()->get(Uuid::fromString($id));
         $category->delete();
+
+        $this->response->redirect('/');
     }
 
     private function getCategoryRepository(): CategoryRepository
     {
         return new CategoryRepository();
+    }
+
+    private function getCategoryAccessChecker(): CategoryAccessChecker
+    {
+        return new CategoryAccessChecker();
     }
 }
